pyproject.toml
[project]
name = "twitter-influencer"
version = "0.1.0"
description = "AI News Scraper with Twitter/X content aggregation and vector similarity search"
requires-python = ">=3.12"
dependencies = [
    "undetected-chromedriver>=3.5.0",
    "python-dotenv>=0.19.0",
    "setuptools>=65.0.0",
    "requests>=2.31.0",
    "sentence-transformers>=2.2.0",
    "FlagEmbedding>=1.2.0",
    "sqlite-vec>=0.1.0",
    "numpy>=1.24.0",
    "scikit-learn>=1.3.0",
    "hdbscan>=0.8.33",
    "pyautogui>=0.9.54",
    "schedule>=1.2.0",
    "pillow>=10.0.0",
    "flask>=3.1.2",
    "youtube-transcript-api>=1.0.0",
    "dateparser>=1.2.0",
    "beautifulsoup4>=4.12.0",
    "feedparser>=6.0.0",
    "opentelemetry-api>=1.20.0",
    "opentelemetry-sdk>=1.20.0",
    "discord.py>=2.3.0",
    "Mastodon.py>=2.0.0",
    "anthropic>=0.40.0",
    "pydantic-ai>=0.1.0",
    "wordninja>=2.0.0",
]

[tool.uv]
package = false
